+++ app/parsers/hyku_addons/csv_parser.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
+
 module HykuAddons
-  # rubocop:disable Metrics/ClassLength
   class CsvParser < Bulkrax::CsvParser
+    include HykuAddons::CollectionBehavior
+
     # FIXME: Override to make debugging easier
     def perform_method
       return :perform_now unless Rails.env.production?
@@ -56,51 +58,11 @@ module HykuAddons
       HykuAddons::CsvAdminSetEntry
     end
 
-    def collections
-      records.map do |r|
-        next unless r[collection_id_key].present?
-
-        collection_ids = r[collection_id_key]&.split(collection_delimiter)
-
-        collections = if r[collection_title_key].present?
-                        collection_titles = r[collection_title_key]&.split(collection_delimiter)
-                        collection_ids.each_with_index.map { |_id, i| { id: collection_ids[i], title: collection_titles[i] } }
-                      else
-                        collection_ids.each_with_index.map { |id, i| { id: id, title: "New collection #{i + 1}" } }
-                      end
-
-        collections
-      end.flatten.compact.uniq
-    end
-
-    # Override to pass Bulkrax.system_identifier_field as singular instead of array
-    def create_collections
-      admin_sets.each_with_index do |admin_set_name, index|
-        next if admin_set_name.blank?
-
-        call_collection_job(admin_set_entry_class, admin_set_name, admin_set_metadata(admin_set_name))
-
-        increment_counters(index, true)
-      end
-
-      collections.each_with_index do |collection, index|
-        next if collection.empty?
-
-        call_collection_job(collection_entry_class, collection[:id], collection_metadata(collection))
-
-        increment_counters(index, true)
-      end
-    end
-
     def admin_sets
       # does the CSV contain an admin_set column?
       return [] unless import_fields.include?(:admin_set)
       # retrieve a list of unique admin sets
       records.map { |r| r[:admin_set] }.flatten.compact.uniq
-    end
-
-    def collections_total
-      collections.size + admin_sets.size
     end
 
     def path_to_files
@@ -148,60 +110,5 @@ module HykuAddons
     rescue StandardError
       @total = 0
     end
-
-    private
-
-      def call_collection_job(item_entry_class, item_id, item_metadata)
-        new_entry = find_or_create_entry(item_entry_class, item_id, 'Bulkrax::Importer', item_metadata)
-
-        begin
-          Bulkrax::ImportWorkCollectionJob.perform_now(new_entry.id, current_run.id)
-        rescue StandardError => e
-          new_entry.status_info(e)
-        end
-      end
-
-      def admin_set_metadata(admin_set)
-        {
-          title: [admin_set],
-          # Allow Hyku to generate a UUID for the admin set
-          # Bulkrax.system_identifier_field => nil,
-          visibility: 'open',
-          collection_type_gid: Hyrax::CollectionType.find_or_create_admin_set_type.gid
-        }
-      end
-
-      def collection_metadata(collection)
-        {
-          title: [collection[:title]],
-          Bulkrax.system_identifier_field => collection[:id],
-          id: collection[:id],
-          visibility: 'open',
-          collection_type_gid: Hyrax::CollectionType.find_or_create_default_collection_type.gid
-        }
-      end
-
-      def collection_delimiter
-        Bulkrax.field_mappings["HykuAddons::CsvParser"]&.dig("collection", :split) || "\|"
-      end
-
-      def collection_prefix
-        if Gem.loaded_specs["bulkrax"].version < Gem::Version.create('3.0')
-          'collection'
-        else
-          'parent'
-        end
-      rescue
-        'collection'
-      end
-
-      def collection_id_key
-        collection_prefix.to_sym
-      end
-
-      def collection_title_key
-        "#{collection_prefix}_title".to_sym
-      end
   end
-  # rubocop:enable Metrics/ClassLength
 end
